+++ packages/titus-fastify-backend/docker/api/Dockerfile.prod
@@ -8,7 +8,7 @@ RUN apk upgrade && apk add postgresql-client
 
 # Copy the server code
 COPY config /titus/config
-COPY theming.js /titus/theming.js
+COPY index.js /titus/index.js
 COPY lib /titus/lib
 COPY node_modules /titus/node_modules
 COPY package.json /titus/package.json
