+++ packages/titus-kitchen-sink-backend/docker/dev.env
@@ -10,7 +10,7 @@ NF_COMMENTS_PGDATABASE=titus
 NF_COMMENTS_PGPASSWORD=titus
 
 # set with the data from your auth0 app
-AUTH0_DOMAIN=
-AUTH0_CLIENT_ID=
-AUTH0_CLIENT_SECRET=
-AUTH0_AUDIENCE=
+# AUTH0_DOMAIN=
+# AUTH0_CLIENT_ID=
+# AUTH0_CLIENT_SECRET=
+# AUTH0_AUDIENCE=
