+++ packages/titus-kitchen-sink/.env
@@ -12,7 +12,7 @@ REACT_APP_COMMENTS_ENDPOINT=ws://localhost:5000
 
 # Set with the data for Auth0 using the OIDC standard
 # REACT_APP_OIDC_AUTHORITY=
-REACT_APP_OIDC_CLIENT_ID=
+# REACT_APP_OIDC_CLIENT_ID=
 
 # Set with the data for your Auth0 app
 # REACT_APP_AUTH0_DOMAIN=
